+++ src/main/java/net/silentchaos512/gear/api/traits/ITraitCondition.java
@@ -1,0 +1,9 @@
+package net.silentchaos512.gear.api.traits;
+
+import net.minecraft.item.ItemStack;
+import net.silentchaos512.gear.api.parts.PartDataList;
+
+@FunctionalInterface
+public interface ITraitCondition {
+    boolean matches(ItemStack gear, PartDataList parts, ITrait trait);
+}